refactor: move tab screens into feature modules
